\name{BASISbinomial}
\alias{BASISbinomial}
\docType{data}
\title{An Example Data File for the Binomial Model}
\description{This is a 500x481 sample feature matrix}
\usage{data(BASISbinomial)}
\format{
  The format is:
 int [1:500, 1:481] 0 -1 0 0 0 0 -1 -1 0 1 ...
}
\details{The data was simulated on a 2400cM chromosome, each column corresponded to an even spaced QTL}
\source{Huang A, Xu S, Cai X: Empirical Bayesian LASSO-logistic regression for multiple binary trait locus mapping. BMC genetics 2013, 14(1):5.}

\examples{
data(BASISbinomial)
}
\keyword{datasets}
