\name{yBinomial}
\alias{yBinomial}
\docType{data}
\title{Sample Variable Data for Binomial Model}
\description{Corresponding to the class label of BASISbinomial}
\usage{data(yBinomial)}
\format{
  The format is:
 int [1:500, 1] 1 1 1 1 1 1 1 1 1 1 ...
}

\source{Huang A, Xu S, Cai X: Empirical Bayesian LASSO-logistic regression for multiple binary trait locus mapping. BMC genetics 2013, 14(1):5.}

\examples{
data(BASISbinomial)
}
\keyword{datasets}
